requirements.txt
sqlalchemy~=2.0.39
pydantic~=2.10.3
fastapi~=0.112.2
pytest~=8.3.4
protobuf~=6.31.0
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.169.0
uliweb-alembic~=0.6.9
python-dotenv~=1.1.0